Revert "fingerprint: goodix: Fix err_[irq|reset] GOTO labels"
If the requested goodix_reset gpio is not found, the first bail out happens
that proceeds to go to err_reset to return rc where rc = 0. This means it returns
0 to kill the function immediately.

If the first gpio request has been satisfied, we move onto the second request in
search of goodix_irq. If that is not found, we need to bail out and return 0. But
first, we need to free up the exisiting gpio that is the successfully requested
goodix_reset. Otherwise, it can cause undefined behaviour. Afterwards, we can
return successfully.

The current GOTO labels are wrong and doesn't do the above mentioned process.

The commit previously has been created due to a misconception. Now that the
misconception has been cleared, revert back to the old state that is actually
correct and discard the faulty logic.

This reverts commit 8ec3559cd6fafd715eb9725ae9d98041c3bfac0e.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>